% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean the Most Recent Dataset}
\usage{
clean_data()
}
\value{
No return value, called for side effects only (creation of two files
  in folder "02 Data")
}
\description{
Clean the Most Recent Dataset
}
