% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create Results Table}
\usage{
create_table(object)
}
\arguments{
\item{object}{An object of class `psythelper`}
}
\value{
A Word file in folder `"04 Figures"`
}
\description{
Create a Word file containing a formated table with all instrument results.
}
