% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_gad_7}
\alias{.evaluate_gad_7}
\title{Evaluate the GAD-7}
\usage{
.evaluate_gad_7(
  data,
  first_variable = .data$x1_nervositaet_aengstlichkeit_oder_anspannung,
  last_variable = .data$x7_gefuehl_der_angst_so_als_wuerde_etwas_schlimmes_passieren
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the GAD-7 questionnaire.
}
