% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_mom_ai}
\alias{.evaluate_mom_ai}
\title{Evaluate the MoM-AI}
\usage{
.evaluate_mom_ai(
  data,
  first_variable = .data$x1_nervoes,
  last_variable = .data$x24_furcht_dass_etwas_schreckliches_geschehen_wird
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the Mind over Mood Anxiety Inventory.
}
