% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_mom_di}
\alias{.evaluate_mom_di}
\title{Evaluate The MoM-DI}
\usage{
.evaluate_mom_di(
  data,
  first_variable = .data$x1_traurige_oder_depressive_stimmung,
  last_variable = .data$x19_verringerter_sexualtrieb
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the Mind over Mood Depression Inventory.
}
