% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_oci_r}
\alias{.evaluate_oci_r}
\title{Evaluate the OCI-R}
\usage{
.evaluate_oci_r(
  data,
  first_variable = .data$x1_ich_bewahre_so_viele_gegenstaende_auf_dass_sie_mich_behindern,
  last_variable = .data$x18_ich_bekomme_haeufig_abscheuliche_gedanken
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the Obsessive-Compulsive Inventory Revised in its German form.
}
