% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_sias}
\alias{.evaluate_sias}
\title{Evaluate the SIAS}
\usage{
.evaluate_sias(
  data,
  first_variable = .data$x1_ich_werde_nervoes,
  last_variable = .data$x20_ich_bin_unsicher_ob
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the Social Interaction Anxiety Scale.
}
