% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_who_5}
\alias{.evaluate_who_5}
\title{Evaluate the WHO-5}
\usage{
.evaluate_who_5(
  data,
  first_variable = .data$x1_war_ich_froh_und_guter_laune,
  last_variable = .data$x5_war_mein_alltag_voller_dinge_die_mich_interessieren
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the World Health Organization well-being index.
}
