% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{.evaluate_wi_d}
\alias{.evaluate_wi_d}
\title{Evaluate the WI-d}
\usage{
.evaluate_wi_d(
  data,
  first_variable = .data$x1_machen_sie_sich_oft_sorgen_krank,
  last_variable = .data$x14_haben_sie_angst_krank_zu_werden
)
}
\arguments{
\item{data}{A dataset containing all responded items for the Instrument in German.}

\item{first_variable}{Name of the first variable}

\item{last_variable}{Name of the last variable}
}
\value{
A tibble containing all items with a sum score
}
\description{
Evaluate the Whiteley Index in its German form.
}
