% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_instruments.R
\name{.ext_instruments}
\alias{.ext_instruments}
\title{Extract Instruments in Dataset}
\usage{
.ext_instruments(data)
}
\arguments{
\item{data}{A tidy dataframe}
}
\value{
A vector of strings
}
\description{
Returns the instruments in a dataframe. It basically detects named columns
with only NAs.
}
