% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_dataset.R
\name{.get_recent_dataset}
\alias{.get_recent_dataset}
\title{Get Most Recent File Dataset in a Folder}
\usage{
.get_recent_dataset(folder_path = "01 Raw Data/")
}
\arguments{
\item{folder_path}{A folder path where the datafiles live. Defaults to "01
Raw Data/"}
}
\value{
A string (filename of most recent file)
}
\description{
Get Most Recent File Dataset in a Folder
}
