% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_fet}
\alias{.plot_fet}
\title{Plot FET}
\usage{
.plot_fet(
  data,
  filename = "FET.png",
  fig_path = "04 Figures/",
  fig_width = 15,
  fig_height = 15,
  fig_units = "cm"
)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}

\item{filename}{File name for saving, defaults to `"FET.png"`}

\item{fig_path}{Path to folder to save to, defaults to `"04 Figures/"`}

\item{fig_width}{Figure width, defaults to 15}

\item{fig_height}{Figure height, defaults to 15}

\item{fig_units}{Dimension units, defaults to `"cm"`}
}
\value{
A ggplot2 object
}
\description{
Plot FET
}
