% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_gad_7}
\alias{.plot_gad_7}
\title{Plot GAD-7}
\usage{
.plot_gad_7(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot GAD-7
}
