% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_instruments}
\alias{.plot_instruments}
\title{Plot Instruments}
\usage{
.plot_instruments(
  data,
  datum = "datum",
  instrument,
  x_label = "Datum",
  y_label = "Instrument",
  y_limit = 0,
  filename = "test.png",
  fig_path = "04 Figures/",
  fig_width = 15,
  fig_height = 10,
  fig_units = "cm"
)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}

\item{datum}{Date column, defaults to "datum"}

\item{instrument}{An instrument column}

\item{x_label}{x-axis label}

\item{y_label}{y-axis label}

\item{y_limit}{Upper y-axis label, defaults to 0}

\item{filename}{File name for saving}

\item{fig_path}{Path to folder to save to, defaults to `"04 Figures/"`}

\item{fig_width}{Figure width, defaults to 15}

\item{fig_height}{Fig height, defaults to 10}

\item{fig_units}{Dimension units, defaults to `"cm"`}
}
\value{
A ggplot2 object
}
\description{
Plot Instruments
}
