% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_mom_ai}
\alias{.plot_mom_ai}
\title{Plot MoM-AI}
\usage{
.plot_mom_ai(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot MoM-AI
}
