% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_mom_di}
\alias{.plot_mom_di}
\title{Plot MoM-DI}
\usage{
.plot_mom_di(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot MoM-DI
}
