% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_oci_r}
\alias{.plot_oci_r}
\title{Plot OCI-R}
\usage{
.plot_oci_r(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot OCI-R
}
