% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_sias}
\alias{.plot_sias}
\title{Plot SIAS}
\usage{
.plot_sias(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot SIAS
}
