% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_success}
\alias{.plot_success}
\title{Plot Therapy success}
\usage{
.plot_success(
  data,
  filename = "Therapieerfolg.png",
  fig_path = "04 Figures/",
  fig_width = 15,
  fig_height = 10,
  fig_units = "cm"
)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}

\item{filename}{File name for savong, defaults to `"Therapieerfolg.png"`}

\item{fig_path}{Path to folder to save to, defaults to `"04 Figures/"`}

\item{fig_width}{Figure width, defaults to 15}

\item{fig_height}{Fig height, defaults to 10}

\item{fig_units}{Dimension units, defaults to `"cm"`}
}
\value{
A ggplot2 object
}
\description{
Plot Therapy success
}
