% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_who_5}
\alias{.plot_who_5}
\title{Plot WHO-5}
\usage{
.plot_who_5(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot WHO-5
}
