% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_instruments.R
\name{.plot_wi_d}
\alias{.plot_wi_d}
\title{Plot WI-d}
\usage{
.plot_wi_d(data)
}
\arguments{
\item{data}{A tidy dataframe containing at least columns "datum" and a
supported instrument}
}
\value{
A ggplot2 object
}
\description{
Plot WI-d
}
