% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{evaluate_instruments}
\alias{evaluate_instruments}
\title{Evaluate All Instruments}
\usage{
evaluate_instruments(data)
}
\arguments{
\item{data}{A dataframe with all responses and named empty columns before
every instrument.}
}
\value{
Does not really return anything, is mainly called for its side
  effects.
}
\description{
Evaluate All Instruments
}
