% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_raw_data.R
\name{ext_raw_data}
\alias{ext_raw_data}
\title{Export Raw Data}
\usage{
ext_raw_data(object)
}
\arguments{
\item{object}{An object of class `psythelper`}
}
\value{
A tibble
}
\description{
After evaluation, export the raw data in R.
}
