% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.psythelper}
\alias{print.psythelper}
\title{Print the results of an Evaluation}
\usage{
\method{print}{psythelper}(x, ...)
}
\arguments{
\item{x}{an object of class `"psythelper"`}

\item{...}{Additional arguments}
}
\value{
Nothing, called for side effects
}
\description{
Print the results of an Evaluation
}
