% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_current_analysis.R
\name{use_current_analysis}
\alias{use_current_analysis}
\title{Use Current Analysis Template}
\usage{
use_current_analysis()
}
\value{
Nothing, used for side effects
}
\description{
Use Current Analysis Template
}
