% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_project.R
\name{use_patient_analysis_template}
\alias{use_patient_analysis_template}
\title{Create an analysis file}
\usage{
use_patient_analysis_template(folder = NA)
}
\arguments{
\item{folder}{Folder path}
}
\value{
Called for side effects.
}
\description{
Create an analysis file
}
