% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_instruments.R
\name{evaluate_instruments}
\alias{evaluate_instruments}
\title{Evaluate All Instruments}
\usage{
evaluate_instruments(data_path = "02 Data/therapie-verlauf.rds")
}
\arguments{
\item{data_path}{Wehre lies the data? Defaults to `"02 Data/therapie-verlauf.rds"`}
}
\value{
Does not really return anything, is mainly called for its side
  effects.
}
\description{
Evaluate All Instruments
}
